feat(frontend): back to wc v1 and installed more wallets
