+++ packages/frontend/package.json
@@ -25,8 +25,10 @@
     "@mui/material": "^5.11.10",
     "@sentry/react": "^7.44.2",
     "@sentry/tracing": "^7.44.2",
+    "@web3-onboard/coinbase": "^2.2.2",
     "@web3-onboard/core": "^2.17.0",
     "@web3-onboard/injected-wallets": "^2.8.4",
+    "@web3-onboard/ledger": "^2.4.4",
     "@web3-onboard/react": "^2.7.2",
     "@web3-onboard/walletconnect": "^2.3.6",
     "@x-fuji/sdk": "0.1.0",
